+++ examples/polymorphic/expected/AllMembers/Members/Organization.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spawnia\Sailor\Polymorphic\AllMembers\Members;
+
+class Organization extends \Spawnia\Sailor\TypedObject
+{
+    /** @var string */
+    public $code;
+
+    public function codeTypeMapper(): callable
+    {
+        return new \Spawnia\Sailor\Mapper\DirectMapper();
+    }
+}