+++ examples/polymorphic/expected/NodeMembers/NodeMembersErrorFreeResult.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spawnia\Sailor\Polymorphic\NodeMembers;
+
+class NodeMembersErrorFreeResult extends \Spawnia\Sailor\ErrorFreeResult
+{
+    public NodeMembers $data;
+}