+++ examples/polymorphic/expected/UserOrPost/UserOrPostErrorFreeResult.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spawnia\Sailor\Polymorphic\UserOrPost;
+
+class UserOrPostErrorFreeResult extends \Spawnia\Sailor\ErrorFreeResult
+{
+    public UserOrPost $data;
+}